tests/wpt/web-platform-tests/navigation-api/navigate-event/intercept-canceled-event.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
test(t => {
  let assertionHappened = false;
  navigation.onnavigate = t.step_func_done(e => {
    e.preventDefault();
    assert_throws_dom("InvalidStateError", () => e.intercept());
    assertionHappened = true;
  });

  location.href = "#1";
  assert_true(assertionHappened);
}, "event.intercept() throws if used on a canceled event");
</script>
